refactor: separate interface IOkCancelRef
