+++ components/AntDesign.csproj
@@ -72,7 +72,14 @@
     <Exec WorkingDirectory="$(SolutionDir)" Command="npm install" />
   </Target>
 
-   
+  <Target Name="DebugRunGulp" BeforeTargets="DebugEnsureNodeEnv" Condition=" '$(Configuration)' == 'Debug' And Exists('$(SolutionDir)node_modules') ">
+    <Exec WorkingDirectory="$(SolutionDir)" Command="npm run gulp:components" />
+  </Target>
+
+  <Target Name="PublishRunGulp" AfterTargets="ComputeFilesToPublish">
+    <Exec WorkingDirectory="$(SolutionDir)" Command="npm install" />
+    <Exec WorkingDirectory="$(SolutionDir)" Command="npm run gulp:components" />
+  </Target>
 
   <ItemGroup>
     <Content Update="**\*.razor">
